+++ neoforge/src/main/java/com/gitlab/srcmc/rctapi/neoforge/client/NeoForgeClient.java
@@ -18,12 +18,14 @@
 package com.gitlab.srcmc.rctapi.neoforge.client;
 
 import com.gitlab.srcmc.rctapi.ModCommon;
+import com.gitlab.srcmc.rctapi.client.ModClient;
+
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.fml.common.Mod;
 
 @Mod(value = ModCommon.MOD_ID, dist = Dist.CLIENT)
 public class NeoForgeClient {
     public NeoForgeClient() {
-        // This entrypoint is suitable for setting up client-specific logic, such as rendering.
+        ModClient.init();
     }
 }